Changed .gitignore and re-initialized game directory
