+++ .gitignore
@@ -8,3 +8,4 @@ __build_time
 .venv
 main
 main.exe
+Example
